Moved logout link to the navbar and removed settings
